pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = ""
authors = [{ name = "mr-reinaldo", email = "josereinaldo.pessoal@gmail.com" }]
readme = "README.md"
requires-python = ">=3.13, <4.0"
dependencies = [
    "fastapi[standard] (>=0.116.1,<0.117.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "pydantic[email,timezone] (>=2.11.7,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "cryptography (>=45.0.5,<46.0.0)",
    "pwdlib[argon2] (>=0.2.1,<0.3.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-cov = "^6.2.1"
taskipy = "^1.14.1"
ruff = "^0.12.3"
aiosqlite = "^0.21.0"

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
run = 'fastapi dev app/main.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=app -vv'
post_test = 'coverage html'
